pinning to same kube client version as servers
