+++ go.mod
@@ -4,29 +4,13 @@ go 1.13
 
 require (
 	github.com/aymerick/raymond v2.0.3-0.20180322193309-b565731e1464+incompatible
-	github.com/davecgh/go-spew v1.1.1
-	github.com/gogo/protobuf v1.2.2-0.20190818160850-4c00d2f19fb9
-	github.com/golang/protobuf v1.3.1
-	github.com/google/gofuzz v1.0.0
-	github.com/googleapis/gnostic v0.0.0-20170729233727-0c5108395e2d
-	github.com/imdario/mergo v0.3.5
-	github.com/json-iterator/go v1.1.7
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd
-	github.com/modern-go/reflect2 v1.0.1
-	github.com/spf13/pflag v1.0.3
-	golang.org/x/crypto v0.0.0-20190605123033-f99c8df09eb5
-	golang.org/x/net v0.0.0-20190812203447-cdfb69ac37fc
-	golang.org/x/oauth2 v0.0.0-20170412232759-a6bd8cefa181
-	golang.org/x/sys v0.0.0-20190606165138-5da285871e9c
-	golang.org/x/text v0.3.2
-	golang.org/x/time v0.0.0-20161028155119-f51c12702a4d
-	google.golang.org/appengine v1.6.2-0.20190801181406-fb139bde60fa
-	gopkg.in/inf.v0 v0.9.0
-	gopkg.in/yaml.v2 v2.2.2
-	k8s.io/api v0.0.0-20190313235455-40a48860b5ab
-	k8s.io/apimachinery v0.0.0-20190313205120-d7deff9243b1
-	k8s.io/client-go v11.0.0+incompatible
-	k8s.io/klog v0.4.0
-	k8s.io/utils v0.0.0-20190221042446-c2654d5206da
-	sigs.k8s.io/yaml v1.1.0
+	github.com/gogo/protobuf v1.2.2-0.20190818160850-4c00d2f19fb9 // indirect
+	github.com/json-iterator/go v1.1.7 // indirect
+	github.com/spf13/pflag v1.0.3 // indirect
+	google.golang.org/appengine v1.6.2-0.20190801181406-fb139bde60fa // indirect
+	gopkg.in/yaml.v2 v2.2.2 // indirect
+	k8s.io/api v0.0.0-20190819141258-3544db3b9e44
+	k8s.io/apimachinery v0.0.0-20190817020851-f2f3a405f61d
+	k8s.io/client-go v0.0.0-20190819141724-e14f31a72a77
+	k8s.io/klog v0.4.0 // indirect
 )
